+++ Marlin/src/inc/Conditionals_LCD.h
@@ -870,6 +870,7 @@
 #if !HAS_LEVELING
   #undef RESTORE_LEVELING_AFTER_G28
   #undef ENABLE_LEVELING_AFTER_G28
+  #undef G29_RETRY_AND_RECOVER
 #endif
 #if !HAS_LEVELING || EITHER(MESH_BED_LEVELING, AUTO_BED_LEVELING_UBL)
   #undef PROBE_MANUALLY
